+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Content loader on Firefox.
 
+### Changed
+- Product Query.
+
 ## [0.1.5] - 2018-08-10
 ### Added
 - `ProductKit` content loader.